Update Python grammar to include keywords introduced in Python 3
This supports https://github.com/microsoft/monaco-editor/issues/1762.

Note the new Python 3 keywords are:

* `async`
* `await`
* `nonlocal`

I also reorganized the list a bit because it seems like it contains
a mixture of keywords and builtins (and `self`, which is neither...),
so it is helpful to be specific to illustrate how to properly maintain
the list.