fix: make all moduleResolution values available
adds additional members to the `ModuleResolutionKind` enum to make all
possible `moduleResolution` values available.

the values are based on the enum in `typescriptServices.js`, namely:
```javascript
      ModuleResolutionKind = /* @__PURE__ */ ((ModuleResolutionKind3) => {
        ModuleResolutionKind3[ModuleResolutionKind3["Classic"] = 1] = "Classic";
        ModuleResolutionKind3[ModuleResolutionKind3["NodeJs"] = 2] = "NodeJs";
        ModuleResolutionKind3[ModuleResolutionKind3["Node10"] = 2] = "Node10";
        ModuleResolutionKind3[ModuleResolutionKind3["Node16"] = 3] = "Node16";
        ModuleResolutionKind3[ModuleResolutionKind3["NodeNext"] = 99] = "NodeNext";
        ModuleResolutionKind3[ModuleResolutionKind3["Bundler"] = 100] = "Bundler";
        return ModuleResolutionKind3;
```

from a developer experience perspective I would actually prefer if the
enum was replaced with the string values that you would use in
`tsconfig.json` as I'd find this more intuitive, but this would be a
larger breaking change.